DHFPROD-4489: Changing mapping step to acceptsBatch=true
This is for user mapping params, and it avoids having to add the new "beforeMain" function. The mapping step main function then iterates over each item. If one throws an error, then it call addStepError and keeps going, unless stopOnError is true. This preserves the existing functionality for a mapping step.