DHFPROD-7288: Handling custom hook errors for 5.5
Doing two things here:

- Ensuring that a Batch doc is written when a hook fails
- Passing the output content array when beforeHook is false

The second thing is being done for two reasons - 1) It's what users likely expected, and 2) It's the behavior users were getting for mapping steps, which is what they were most likely attaching custom hooks to for the purpose of modifying URIs. Mapping steps happened to allow this because they were modifying the input content object instead of returning a new one. But now in 5.5, a mapping step can be returning new content objects since it can return multiple content objects for related entities. So changing this both provides the expected behavior for an "after" hook and also ensures that hooks will continue to work as expected for mapping steps.