DHFPROD-7288: Batch status is now correct when custom hook fails
Was missing a call to stepExecutionContext.addBatchErrorForEntireBatch . 

Also, for sanity reasons, removed the datahub instance from flow.sjs, as it had almost no value and created a circular dependency. Removed the code that returned the performance metrics, but that's currently disabled anyway in stepDefinition.sjs. If we find we have a need for that at some point, I think we'll integrate it in a different fashion.