DHFPROD-7334: Fixed staging rewriter; no longer forking REST modules
Deleted all of the forked REST modules, along with 4 amps that are no longer required.

Renamed amp files for consistency (this has no bearing on how amps are deployed). 

Removed a bunch of Java code for determining the staging rewriter to use based on server version. That's no longer needed now that the staging rewriter is generated dynamically when DHF modules are loaded. 

Added tests for ml:hubversion, ml:flow, and ml:sjsFlow to ensure they work.