DHFPROD-7570: stopOnError now works when the collector fails
I had to update a couple existing tests, which I think in hindsight were erroneously expecting the wrong job status. If stopOnError=true, and the collector fails, then a user expects the flow to be stopped and no other steps to execute, which means the job status should be stop-on-error.