DHFPROD-7629: Upgraded to ml-app-deployer / ml-gradle 4.3.0
Updated other dependencies to match those of Java Client 5.5.0.

One exception is in HC - I updated Spring Boot to 2.5.3 and Spring Security to 5.5.1, which aligns those with the version of Spring used by ml-app-deployer. But the former uses Jackson 2.12.x, while the Java Client uses 2.11.x. This does not appear to be an issue though - all the HC tests run fine. 

As part of the Spring Boot upgrade, made two little changes based on SB release notes. Removed the deprecated-now-removed "getErrorPath" method in favor of server.error.path (though that appears unused in HC). And changed setAllowedOrigins to setAllowedOriginPatterns.

Also bumped up the aws-java-sdk dependency in HC, as HC was on a fairly old version with older versions of Jackson.

Removed jcenter() as a repository from the main build files (cleaning up the example projects will be a separate effort).