DHFPROD-7663: Can now run DHF on Gradle 7
Most changes are to ml-data-hub tests and the example projects; Gradle 7 uses Groovy 3, so had to bump up spock test to 2.0-3.0. Then had to modify many tests to not use the junit4 TemporaryFolder class, but rather a plain File.

For example projects - bumped them to gradlew 7.1.1, and replaced jcenter with mavenCentral.

Fixed several DHF tasks; task properties apparently should not have "public" on them, because otherwise you need to include a public getter. So it's simpler / better just to have e.g. "String flowName" with @Input and @Optional on it. Also fixed every other task property that didn't already have @Input on it, or added @Internal if appropriate. 

DHF will now also scaffold out a gradlew properties file for Gradle 7.1.1.