DHFPROD-7673: Removed extract-document-data from search options
This wasn't being used by HC at all. Also removed occurrences of "extracted" from UI mock data. And removed AbstractModel and its assertion, which was based on looking for the word e.g. "Customer" in the search options, but it was only there because extract-document-data happened to have it. That was a fairly weak assertion, so removed it.