DHFPROD-7700: Removing obsolete code from EntityManager
This code is made obsolete by removal of the undocumented "entityName" parameter in the undocumented "hubCreateMapping" Gradle task. Most of the EntityManager code could be removed without removing the "entityName" bit, but because it's not part of the public API (can only know about it by looking at the code), and the value is so low, we're removing it and all the related EntityManager code, and some MappingManager code too that is dead. 

Also renamed "EntityServicesAlignmentTest" to "GenerateTdeTest".