DHFPROD-7725: Explore now shows properties for XML entities
The code in Detail.tsx wasn't handling XML docs correctly. Was able to fix this in the backend via getEntityInstanceProperties, and as a result, was able to delete a good chunk of code from both the UI and the backend. 

Changed some functions in entity-search-lib to require a doc instead of a URI, which means we have fewer calls to cts.doc(uri) in the application code. 

Also fixed a test performance issue by renaming setup.xqy to suite-setup.xqy, as the content only needs to be loaded once for all the tests in that suite.