DHFPROD-7807: Not ignoring entity-config/databases
This ensures that the databases files will be accounted for when mlDeploy is run. Without this, a user must run "mlDeploy hubDeployAsDeveloper", as the latter does generate index files while the former does not. One reason that the former does not is a chicken-and-egg problem - generating the indexes file requires first loading the DHF modules, but the databases/indexes are deployed before DHF modules are loaded. This can be solved, but the easiest solution is just to not gitignore src/main/entity-config/databases.